add tests for rate limiter routes
