+++ test/rateLimiter.test.ts
@@ -7,19 +7,6 @@ describe('[nuxt-security] Rate Limiter', async () => {
     rootDir: fileURLToPath(new URL('./fixtures/rateLimiter', import.meta.url)),
   })
 
-  it ('should return 200 OK after multiple requests for certain route', async () => {
-    const res1 = await fetch('/test')
-    await fetch('/test')
-    await fetch('/test')
-    await fetch('/test')
-    const res5 = await fetch('/test')
-
-    expect(res1).toBeDefined()
-    expect(res1).toBeTruthy()
-    expect(res5.status).toBe(200)
-    expect(res5.statusText).toBe('OK')
-  })
-
   it ('should return 200 OK when not reaching the limit', async () => {
     const res1 = await fetch('/')
     const res2 = await fetch('/')
@@ -41,5 +28,39 @@ describe('[nuxt-security] Rate Limiter', async () => {
     expect(res1).toBeTruthy()
     expect(res5.status).toBe(429)
     expect(res5.statusText).toBe('Too Many Requests')
+  })
+
+  it ('should return 200 OK after multiple requests for a route with a higher limit', async () => {
+    const res1 = await fetch('/test')
+    await fetch('/test')
+    await fetch('/test')
+    await fetch('/test')
+    await fetch('/test')
+    await fetch('/test')
+    const res5 = await fetch('/test')
+
+    expect(res1).toBeDefined()
+    expect(res1).toBeTruthy()
+    expect(res5.status).toBe(200)
+    expect(res5.statusText).toBe('OK')
+  })
+
+  it ('should return 429 when limit reached for a route, but 200 for another route with a higher limit', async () => {
+    const res1 = await fetch('/')
+    await fetch('/')
+    await fetch('/')
+    await fetch('/')
+    const res5 = await fetch('/')
+
+    expect(res1).toBeDefined()
+    expect(res1).toBeTruthy()
+    expect(res5.status).toBe(429)
+    expect(res5.statusText).toBe('Too Many Requests')
+
+    const res6 = await fetch('/test')
+    expect(res6).toBeDefined()
+    expect(res6).toBeTruthy()
+    expect(res6.status).toBe(200)
+    expect(res6.statusText).toBe('OK')
   })
 })
